BSParams-class {BSgenome}	R Documentation

Class "BSParams"

Description

A parameter class for representing all parameters needed for running the bsapply method.

Objects from the Class

Objects can be created by calls of the form new("BSParams", ...).

Slots

X:

a BSgenome object that contains chromosomes that you wish to apply FUN on

FUN:

the function to apply to each chromosome in the BSgenome object 'X'

exclude:

this is a character vector with strings that will be treated as regular expressions to filter out chromosomes whose names match these strings.

simplify:

TRUE/FALSE value to indicate whether or not the function should try to simplify the output for you.

maskList:

A named logical vector of maskStates preferred when used with a BSGenome object. When using the bsapply function, the masks will be set to the states in this vector.

motifList:

A character vector which should contain motifs that the user wishes to mask from the sequence.

userMask:

A IntegerRangesList object, where each element masks the corresponding chromosome in X. This allows the user to conveniently apply masks besides those included in X.

invertUserMask:

A logical indicating whether to invert each mask in userMask.

Methods

bsapply(p)

Performs the function FUN using the parameters contained within BSParams.

Author(s)

Marc Carlson

See Also

bsapply

[Package BSgenome version 1.58.0 Index]